updated app initialization
